update lua quyen khi tao tai khoan,cap nhat quyen,vv

+++ src/main/java/com/library/librarymanager/repository/ChucNangRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ChucNangRepository
-        extends JpaRepository<ChucNang, Long> {
+        extends JpaRepository<ChucNang, Integer> {
     List<ChucNang> findByTenChucNangStartingWith(String prefix);
 
 }
